Move the icons to the images folder, unify KirokuIcons and KirokuImages
