+++ assets/images/delete-session.png
