+++ assets/images/strong-shot.png
