+++ src/components/Icon/KirokuIcons.ts
@@ -1,47 +1,53 @@
-import Achievements from '@assets/icons/achievements.png';
-import AddImage from '@assets/icons/add-image.png';
-import AddUser from '@assets/icons/add-user.png';
-import AlcoholAssortment from '@assets/icons/alcohol-assortment.png';
-import ArrowBack from '@assets/icons/arrow-back.png';
-import ArrowDown from '@assets/icons/arrow-down.png';
-import ArrowRight from '@assets/icons/arrow-right.svg';
-import ArrowUp from '@assets/icons/arrow-up.svg';
-import ArrowUpDown from '@assets/icons/arrows-updown.svg';
-import BarMenu from '@assets/icons/bar-menu.png';
-import Beer from '@assets/icons/beer.png';
-import Bell from '@assets/icons/bell.png';
-import Book from '@assets/icons/book.png';
-import Bug from '@assets/icons/bug.png';
-import Camera from '@assets/icons/camera.png';
-import Check from '@assets/icons/check.png';
-import Checkmark from '@assets/icons/checkmark.svg';
-import Close from '@assets/icons/close.png';
-import Cocktail from '@assets/icons/cocktail.png';
-import Copy from '@assets/icons/copy.png';
-import DeleteSession from '@assets/icons/delete-session.png';
-import Delete from '@assets/icons/delete.png';
-import DotIndicator from '@assets/icons/dot-indicator.svg';
-import Download from '@assets/icons/download.png';
-import Edit from '@assets/icons/edit.png';
-import Exit from '@assets/icons/exit.png';
-import FriendList from '@assets/icons/friend-list.png';
-import Idea from '@assets/icons/idea.png';
-import Menu from '@assets/icons/menu.png';
-import Minus from '@assets/icons/minus.png';
-import Pencil from '@assets/icons/pencil.png';
-import Plus from '@assets/icons/plus.png';
-import RemoveUser from '@assets/icons/remove-user.png';
-import Remove from '@assets/icons/remove.png';
-import Search from '@assets/icons/search.png';
-import Settings from '@assets/icons/settings.png';
-import Share from '@assets/icons/share.png';
-import Social from '@assets/icons/social.png';
-import Statistics from '@assets/icons/statistics.png';
-import StrongShot from '@assets/icons/strong-shot.png';
-import ThinX from '@assets/icons/thin-x.png';
-import ThreeDots from '@assets/icons/three-dots.svg';
-import WeakShot from '@assets/icons/weak-shot.png';
-import Wine from '@assets/icons/wine.png';
+import Achievements from '@assets/images/achievements.png';
+import AddImage from '@assets/images/add-image.png';
+import AddUser from '@assets/images/add-user.png';
+import AlcoholAssortment from '@assets/images/alcohol-assortment.png';
+import ArrowBack from '@assets/images/arrow-back.png';
+import ArrowDown from '@assets/images/arrow-down.png';
+import ArrowRight from '@assets/images/arrow-right.svg';
+import ArrowUp from '@assets/images/arrow-up.svg';
+import ArrowUpDown from '@assets/images/arrows-updown.svg';
+import BarMenu from '@assets/images/bar-menu.png';
+import Beer from '@assets/images/beer.png';
+import Bell from '@assets/images/bell.png';
+import Book from '@assets/images/book.png';
+import Bug from '@assets/images/bug.png';
+import Camera from '@assets/images/camera.png';
+import Check from '@assets/images/check.png';
+import Checkmark from '@assets/images/checkmark.svg';
+import Close from '@assets/images/close.png';
+import Cocktail from '@assets/images/cocktail.png';
+import Copy from '@assets/images/copy.png';
+import Delete from '@assets/images/delete.png';
+import DeleteSession from '@assets/images/delete-session.png';
+import DotIndicator from '@assets/images/dot-indicator.svg';
+import Download from '@assets/images/download.png';
+import Edit from '@assets/images/edit.png';
+import Exit from '@assets/images/exit.png';
+import FriendList from '@assets/images/friend-list.png';
+import Idea from '@assets/images/idea.png';
+import Logo from '@assets/images/app-logo.png';
+import Menu from '@assets/images/menu.png';
+import Minus from '@assets/images/minus.png';
+import Pencil from '@assets/images/pencil.png';
+import Plus from '@assets/images/plus.png';
+import QrCodeFullColor from '@assets/images/kiroku-qr-code-full-color.png';
+import QrCodePlain from '@assets/images/kiroku-qr-code-plain.png';
+import QrCodeWithLogo from '@assets/images/kiroku-qr-code-with-logo.png';
+import Remove from '@assets/images/remove.png';
+import RemoveUser from '@assets/images/remove-user.png';
+import Search from '@assets/images/search.png';
+import Settings from '@assets/images/settings.png';
+import Share from '@assets/images/share.png';
+import Social from '@assets/images/social.png';
+import Statistics from '@assets/images/statistics.png';
+import StrongShot from '@assets/images/strong-shot.png';
+import ThinX from '@assets/images/thin-x.png';
+import ThreeDots from '@assets/images/three-dots.svg';
+import UnderMaintenance from '@assets/images/under-maintenance.jpg';
+import UserIcon from '@assets/temp/user-icon.png';
+import WeakShot from '@assets/images/weak-shot.png';
+import Wine from '@assets/images/wine.png';
 
 export {
   Achievements,
@@ -64,20 +70,24 @@ export {
   Close,
   Cocktail,
   Copy,
-  DeleteSession,
   Delete,
-  Download,
+  DeleteSession,
   DotIndicator,
+  Download,
   Edit,
   Exit,
   FriendList,
   Idea,
+  Logo,
   Menu,
   Minus,
   Pencil,
   Plus,
-  RemoveUser,
+  QrCodeFullColor,
+  QrCodePlain,
+  QrCodeWithLogo,
   Remove,
+  RemoveUser,
   Search,
   Settings,
   Share,
@@ -86,6 +96,8 @@ export {
   StrongShot,
   ThinX,
   ThreeDots,
+  UnderMaintenance,
+  UserIcon,
   WeakShot,
   Wine,
 };
