--- src/components/Icon/KirokuImages.ts
@@ -1,15 +1,0 @@
-import Logo from '@assets/images/app-logo.png';
-import UnderMaintenance from '@assets/images/under-maintenance.jpg';
-import UserIcon from '@assets/temp/user-icon.png';
-import QrCodePlain from '@assets/images/kiroku-qr-code-plain.png';
-import QrCodeWithLogo from '@assets/images/kiroku-qr-code-with-logo.png';
-import QrCodeFullColor from '@assets/images/kiroku-qr-code-full-color.png';
-
-export {
-  Logo,
-  UnderMaintenance,
-  UserIcon,
-  QrCodePlain,
-  QrCodeWithLogo,
-  QrCodeFullColor,
-};
