+++ src/libs/UserUtils.ts
@@ -2,7 +2,7 @@
 import type {OnyxEntry} from 'react-native-onyx';
 import type {ValueOf} from 'type-fest';
 import * as defaultAvatars from '@components/Icon/DefaultAvatars';
-import {UserIcon} from '@components/Icon/KirokuImages';
+import {UserIcon} from '@components/Icon/KirokuIcons';
 import CONST from '@src/CONST';
 import type {LoginList} from '@src/types/onyx';
 import type {UserID} from '@src/types/onyx/OnyxCommon';
